requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
numpy==2.1.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
seaborn==0.13.2
tabulate==0.9.0
pillow==11.0.0
matplotlib==3.9.2